add in_json with MIT jsmn WIP
as a submodule

+++ src/in_dxf.c
@@ -684,7 +684,7 @@ dwg_indxf_variable_type(Dwg_Data * dwg, Bit_Chain *dat, Dwg_Object* obj)
   int is_entity;
 
   i = obj->type - 500;
-  if (i < 0 || i >= dwg->num_classes)
+  if (i < 0 || i >= (int)dwg->num_classes)
     return DWG_ERR_INVALIDTYPE;
 
   klass = &dwg->dwg_class[i];
